Extract AbstractActivationAware from AbstractNode.
